+++ lib/state/chart.dart
@@ -1,5 +1,4 @@
 import 'dart:developer' as developer;
-import 'dart:html';
 
 import 'package:flutter/material.dart';
 import 'package:rubin_chart/rubin_chart.dart';
